+++ tests/e2e/cases/dynamic_require_static_buffer_search/bsearch_4.ts
@@ -1,0 +1,1 @@
+module.exports.label = "bsearch_4";